+++ TrueCraft.Client/Rendering/Blocks/SnowRenderer.cs
@@ -25,7 +25,7 @@ namespace TrueCraft.Client.Rendering.Blocks
         };
 
         public override Vertex[] Render(BlockDescriptor descriptor, Vector3 offset,
-             Tuple<int, int> textureMap, int indiciesOffset, out ushort[] indicies)
+             Tuple<int, int> textureMap, int indiciesOffset, out uint[] indicies)
         {
             var overhead = new Vector3(0.5f, 0.5f, 0.5f);
             var cube = CreateUniformCube(overhead, Texture, indiciesOffset, out indicies, Color.White);
